src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=src/requirements.txt src/requirements/requirements.in
#
asgiref==3.7.2
    # via
    #   django
    #   django-cors-headers
    #   django-countries
asttokens==2.4.1
    # via icecream
autopep8==2.0.4
    # via -r src/requirements/requirements.in
certifi==2024.2.2
    # via httpx
cffi==1.16.0
    # via cryptography
chardet==3.0.4
    # via httpx
colorama==0.4.6
    # via icecream
cryptography==42.0.3
    # via -r src/requirements/requirements.in
dj-database-url==2.1.0
    # via -r src/requirements/requirements.in
django==5.0.2
    # via
    #   -r src/requirements/requirements.in
    #   dj-database-url
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   django-modeltranslation
    #   django-tinymce
    #   djangorestframework
django-cors-headers==4.3.1
    # via -r src/requirements/requirements.in
django-countries==7.5.1
    # via -r src/requirements/requirements.in
django-extensions==3.2.3
    # via -r src/requirements/requirements.in
django-filter==23.5
    # via -r src/requirements/requirements.in
django-ipware==6.0.4
    # via -r src/requirements/requirements.in
django-modeltranslation==0.18.11
    # via -r src/requirements/requirements.in
django-split-settings==1.2.0
    # via -r src/requirements/requirements.in
django-tinymce==3.7.1
    # via -r src/requirements/requirements.in
djangorestframework==3.14.0
    # via -r src/requirements/requirements.in
executing==2.0.1
    # via icecream
googletrans==3.0.0
    # via -r src/requirements/requirements.in
gunicorn==21.2.0
    # via -r src/requirements/requirements.in
h11==0.9.0
    # via httpcore
h2==3.2.0
    # via httpcore
hpack==3.0.0
    # via h2
hstspreload==2024.3.1
    # via httpx
httpcore==0.9.1
    # via httpx
httpx==0.13.3
    # via googletrans
hyperframe==5.2.0
    # via h2
icecream==2.1.3
    # via -r src/requirements/requirements.in
idna==2.10
    # via httpx
packaging==23.2
    # via gunicorn
pillow==10.2.0
    # via -r src/requirements/requirements.in
psycopg2==2.9.9
    # via -r src/requirements/requirements.in
psycopg2-binary==2.9.9
    # via -r src/requirements/requirements.in
pycodestyle==2.11.1
    # via autopep8
pycparser==2.21
    # via cffi
pygeoip==0.3.2
    # via -r src/requirements/requirements.in
pygments==2.17.2
    # via icecream
python-decouple==3.8
    # via -r src/requirements/requirements.in
python-ipware==2.0.2
    # via django-ipware
pytz==2024.1
    # via
    #   -r src/requirements/requirements.in
    #   djangorestframework
rfc3986==1.5.0
    # via httpx
six==1.16.0
    # via
    #   -r src/requirements/requirements.in
    #   asttokens
sniffio==1.3.1
    # via
    #   httpcore
    #   httpx
sqlparse==0.4.4
    # via django
typing-extensions==4.9.0
    # via
    #   dj-database-url
    #   django-countries
    #   django-modeltranslation
tzdata==2024.1
    # via
    #   django
    #   tzlocal
tzlocal==5.2
    # via -r src/requirements/requirements.in
